added some comments and made code easier to read
