+++ client/src/App.tsx
@@ -8,7 +8,7 @@ import VideoNotes from "./routes/VideoNotes";
 import PdfNotes from "./routes/PdfNotes";
 import SchedulePresentation from "./routes/SchedulePresentation";
 import Presentations from "./routes/Presentations";
-import PresentationRoomTest from "./routes/PresentationRoomTest";
+import PresentationRoom from "./routes/PresentationRoom";
 
 
 export default function App() {
@@ -23,8 +23,7 @@ export default function App() {
 				<Route path="/pdf" element={<PdfNotes />} />
 				<Route path="/presentations" element={<Presentations />} />
 				<Route path="/schedulepresentation" element={<SchedulePresentation />} />
-				<Route path="/room/:id" element={<PresentationRoomTest />} />
-			
+				<Route path="/room/:id" element={<PresentationRoom />} />
 			</Routes>
 		</BrowserRouter>
 	);
